Fix blur example usage of clGetDeviceInfo (#136)
`blur` example is retrieving `CL_DEVICE_OPENCL_C_VERSION` in a fixed
sized array of 64 bytes, CL spec does not specify a limit to the output
length. Therefore, the test may fail if the version is longer than 64
bytes.

This change fixes the issue by querying `clGetDeviceInfo` for the
expected output size.

Signed-off-by: Michael Rizkalla <michael.rizkalla@arm.com>